+++ apps/client/src/app/pages/resources/guides/resources-guides.component.ts
@@ -1,7 +1,8 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
 @Component({
+  changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [RouterModule],
   selector: 'gf-resources-guides',
   styleUrls: ['./resources-guides.component.scss'],